* Moved "Ringer" base class to collections package.
