+++ ringer/collections/__init__.py
@@ -1,2 +1,3 @@
+from .base import Ringer
 from .deques import RingerDeque
 from .lists import RingerList
