+++ ringer/collections/base.py
